SUW: Create minimal setup wizard
Brings the Setup Wizard down to the absolute minimum required in order to set up
a device and leave it in a provisioned state (which isn't always possible if the
Setup Wizard is simply omitted). Essentially, this leaves only the following:

- Bluetooth setup (which doesn't require user interaction)
- Welcome page
- Date/time
- Device-specific settings
- Theme (light/dark)
- Navigation bar style
- The final page that shows the logo and start button

All of the other settings from Setup Wizard (i.e. Wi-Fi, SIM, etc.) can already
be configured in Settings, so this allows the initial setup to be streamlined
and let the user "get right into it."

Thanks to XDA user Aqq123, whose work on modifying the Setup Wizard script was
the inspiration for this patch.